Remove schema export from API route
