+++ apps/web/app/api/bug/route.ts
@@ -4,9 +4,10 @@ import { routeHandlerRatelimiter } from "@/lib/rate-limiter";
 import Bowser from "bowser";
 import { type NextRequest, NextResponse } from "next/server";
 import { Resend } from "resend";
-import { z } from "zod";
 
-const createResendClient = () => new Resend(process.env.RESEND_API_KEY);
+import { BugReportSchema } from "./BugReportSchema";
+
+const resend = new Resend(process.env.RESEND_API_KEY);
 
 const parseToJSON = (value: string) => {
   try {
@@ -15,28 +16,6 @@ const parseToJSON = (value: string) => {
     return value;
   }
 };
-
-const BugReportNameSchema = z.string().min(1, "Provide your name");
-const BugReportEmailSchema = z
-  .string()
-  .min(1, "Provide your email")
-  .email("Provide a valid email address");
-const BugReportDescriptionSchema = z
-  .string()
-  .trim()
-  .min(1, "Provide a description");
-const BugReportUserAgentSchema = z.string().min(1, "Provide your user agent");
-const BugReportAdditionalInformationSchema = z.string();
-
-export const BugReportSchema = z.object({
-  name: BugReportNameSchema,
-  email: BugReportEmailSchema,
-  description: BugReportDescriptionSchema,
-  userAgent: BugReportUserAgentSchema,
-  additionalInformation: BugReportAdditionalInformationSchema,
-});
-
-export type BugReport = z.infer<typeof BugReportSchema>;
 
 export const POST = async (req: NextRequest) => {
   const maybeRateLimitedResponse = await routeHandlerRatelimiter(req);
@@ -77,8 +56,6 @@ export const POST = async (req: NextRequest) => {
       status: 403,
     });
   }
-
-  const resend = createResendClient();
 
   const response = await resend.emails.send({
     from: `${name} <${SENDER.email}>`,
